950 Z0Rpt and rename Z1->Z3
